feat: Introduce generics to Calculator
